\name{exonsByGene}
\docType{data}
\alias{exonsByGene}
\alias{exonsByTranscript}
\title{Subset of Ensembl genes/transcripts on chromosome 1}
\description{
\code{exonsByGene} is a \code{GRangesList} object with a subset of
genes/transcripts from the GRCh37 Ensembl annotations. For complete
details on the creation of this object, please see the vignette.
}
\usage{
data("exonsByGene")
data("exonsByTranscript")
}
\format{GRangesList}
\source{Ensembl}
\keyword{datasets}
